refactor: menu패키지로 product클래스 이동
