+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -1,10 +1,10 @@
 package kitchenpos.application;
 
-import kitchenpos.product.application.ProductService;
-import kitchenpos.product.application.dto.ProductCreateRequest;
-import kitchenpos.product.application.dto.ProductResponse;
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.repository.ProductRepository;
+import kitchenpos.menu.application.ProductService;
+import kitchenpos.menu.application.dto.ProductCreateRequest;
+import kitchenpos.menu.application.dto.ProductResponse;
+import kitchenpos.menu.domain.Product;
+import kitchenpos.menu.domain.repository.ProductRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
