+++ src/test/java/kitchenpos/domain/ProductTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain;
 
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
